% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PakPMICS2018wm.R
\docType{data}
\name{PakPMICS2018wm}
\alias{PakPMICS2018wm}
\title{Multiple Indicator Cluster Survey (MICS) 2018 Child Questionnaire Data for Punjab, Pakistan}
\format{A data frame with 157899 rows and 62 variables:
\describe{
   \item{\code{HH1}}{Cluster number}
   \item{\code{HH2}}{Household number}
   \item{\code{WM1}}{Cluster number}
   \item{\code{WM2}}{Household number}
   \item{\code{WMINT}}{Interviewer number}
   \item{\code{WM3}}{Woman's line number}
   \item{\code{LN}}{Line number}
   \item{\code{WM4}}{Supervisor's number}
   \item{\code{WM5}}{Interviewer's number}
   \item{\code{WM6D}}{Day of interview}
   \item{\code{WM6M}}{Month of interview}
   \item{\code{WM6Y}}{Year of interview}
   \item{\code{WM8}}{Household interviewer}
   \item{\code{WM9}}{Consent}
   \item{\code{WM17}}{Result of woman's interview}
   \item{\code{WM7H}}{Start of interview - Hour}
   \item{\code{WM7M}}{Start of interview - Minutes}
   \item{\code{WM10H}}{End of interview - Hour}
   \item{\code{WM10M}}{End of interview - Minutes}
   \item{\code{WM11}}{Interview completed in private}
   \item{\code{WM12}}{Language of the Questionnaire}
   \item{\code{WM13}}{Language of the Interview}
   \item{\code{WM14}}{Native language of the Respondent}
   \item{\code{WM15}}{Translator used}
   \item{\code{WMHINT}}{Household interviewer}
   \item{\code{INTROBH}}{Introduction Birth History (BH)}
   \item{\code{INTROMM}}{Introduction MM Listing}
   \item{\code{MMAUX}}{Other brother/sisters}
   \item{\code{WB3M}}{Month of birth of woman}
   \item{\code{WB3Y}}{Year of birth of woman}
   \item{\code{WB4}}{Age of woman}
   \item{\code{WB5}}{Ever attended school}
   \item{\code{WB6A}}{Highest level of school attended}
   \item{\code{WB6B}}{Highest grade attended at that level}
   \item{\code{WB7}}{Grade completion}
   \item{\code{WB9}}{Attended school during current school year}
   \item{\code{WB10A}}{Level of education attended current school year}
   \item{\code{WB10B}}{Grade of education attended current school year}
   \item{\code{WB11}}{Attended school previous school year}
   \item{\code{WB12A}}{Level of education attended previous school year}
   \item{\code{WB12B}}{Grade of education attended previous school year}
   \item{\code{WB14}}{Can read part of the sentence}
   \item{\code{WB15}}{Duration of living in current place}
   \item{\code{WB16}}{Place of living prior to moving to current place}
   \item{\code{WB17}}{Region prior to moving to current place}
   \item{\code{WB18}}{Covered by health insurance}
   \item{\code{WB19A}}{Type of health insurance: Public health organization}
   \item{\code{WB19B}}{Type of health insurance: Health insurance through employer}
   \item{\code{WB19C}}{Type of health insurance: Social security}
   \item{\code{WB19D}}{Type of health insurance: Other privately purchased commercial health insurance}
   \item{\code{WB19X}}{Type of health insurance: Other}
   \item{\code{WB19NR}}{Type of health insurance: No response}
   \item{\code{WDOI}}{Date of interview women (CMC)}
   \item{\code{WAGE}}{Age}
   \item{\code{MT1}}{Frequency of reading newspaper or magazine}
   \item{\code{MT2}}{Frequency of listening to the radio}
   \item{\code{MT3}}{Frequency of watching TV}
   \item{\code{MT4}}{Ever used a computer or a tablet}
   \item{\code{MT5}}{Computer / tablet usage in the last 3 months}
   \item{\code{MT6A}}{During the last 3 months: Copy or move a file or folder}
   \item{\code{MT6B}}{During the last 3 months: Use a copy / paste in document}
   \item{\code{MT6C}}{During the last 3 months: Send e-mail with attached file}
   \item{\code{MT6D}}{During the last 3 months: Use a basic arithmetic formula in a  spreadsheet}
   \item{\code{MT6E}}{During the last 3 months: Connect and install a new device}
   \item{\code{MT6F}}{During the last 3 months: Install and configure software}
   \item{\code{MT6G}}{During the last 3 months: Create an electronic presentation}
   \item{\code{MT6H}}{During the last 3 months: Transfer a file}
   \item{\code{MT6I}}{During the last 3 months: Write a computer program}
   \item{\code{MT9}}{Ever used internet}
   \item{\code{MT10}}{Internet usage in the last 3 months}
   \item{\code{MT11}}{Own a mobile phone}
   \item{\code{MT12}}{Mobile phone usage in the last 3 months}
   \item{\code{MA1}}{Currently married}
   \item{\code{MA2}}{Age of husband}
   \item{\code{MA3}}{Husband has more wives}
   \item{\code{MA4}}{Number of other wives or partners}
   \item{\code{MA5}}{Ever married}
   \item{\code{MA6}}{Marital status}
   \item{\code{MA7}}{Married or lived with a man once or more than once}
   \item{\code{MA8M}}{Month of first marriage}
   \item{\code{MA8Y}}{Year of first marriage}
   \item{\code{MA11}}{Age at first marriage}
   \item{\code{CM1}}{Ever given birth}
   \item{\code{CM2}}{Any sons or daughters living with you}
   \item{\code{CM3}}{Sons living with you}
   \item{\code{CM4}}{Daughters living with you}
   \item{\code{CM5}}{Any sons or daughters not living with you}
   \item{\code{CM6}}{Sons living elsewhere}
   \item{\code{CM7}}{Daughters living elsewhere}
   \item{\code{CM8}}{Ever had child who later died}
   \item{\code{CM9}}{Boys dead}
   \item{\code{CM10}}{Girls dead}
   \item{\code{CM11}}{Children ever born}
   \item{\code{CM12}}{Confirm total number of ever born}
   \item{\code{CM15}}{Check total number of children ever born}
   \item{\code{CM17}}{Live births in last two years}
   \item{\code{BH11}}{Any other live births}
   \item{\code{DB2}}{Wanted last child then}
   \item{\code{DB4}}{Wanted child later or did not want more children}
   \item{\code{MN2}}{Received prenatal care}
   \item{\code{MN3A}}{Prenatal care provider: Doctor}
   \item{\code{MN3B}}{Prenatal care provider: Nurse/Midwife}
   \item{\code{MN3C}}{Prenatal care provider: Lady Health Visitor (LHV)}
   \item{\code{MN3D}}{Prenatal care provider: Community Mid Wife}
   \item{\code{MN3F}}{Prenatal care provider: Traditional birth attendant}
   \item{\code{MN3X}}{Prenatal care provider: Other}
   \item{\code{MN3NR}}{Prenatal care provider: No Response}
   \item{\code{MN4AU}}{Weeks or months pregnent at first prenatal care - unit}
   \item{\code{MN4AN}}{Weeks or months pregnant at first prenatal care - number}
   \item{\code{MN5}}{Times received prenatal care}
   \item{\code{MN6A}}{Blood pressure}
   \item{\code{MN6B}}{Urine sample}
   \item{\code{MN6C}}{Blood sample}
   \item{\code{MN6D}}{Weight}
   \item{\code{MN6E}}{Importance of Spacing}
   \item{\code{MN6F}}{Information provided for FP methods}
   \item{\code{MN7}}{Has own immunization card}
   \item{\code{MN8}}{Any tetanus toxoid injection during last pregnancy}
   \item{\code{MN9}}{Doses of tetanus toxoid during last pregnancy}
   \item{\code{MN11}}{Any tetanus toxoid injection before last pregnancy}
   \item{\code{MN12}}{Doses of tetanus toxoid before last pregnancy}
   \item{\code{MN14}}{Years ago last tetanus toxoid received before last pregnancy}
   \item{\code{MN16}}{Any SP/Fansidar during last pregnancy}
   \item{\code{MN17}}{Doses of SP/Fansidar during last pregnancy}
   \item{\code{MN18A}}{Received SP/Fansidar during an antenatal care visit, or during another visit: Antenatal visit}
   \item{\code{MN18B}}{Received SP/Fansidar during an antenatal care visit, or during another visit: Another facility visit}
   \item{\code{MN18X}}{Received SP/Fansidar during an antenatal care visit, or during another visit: Other source}
   \item{\code{MN18NR}}{Received SP/Fansidar during an antenatal care visit, or during another visit: No response}
   \item{\code{MN19A}}{Assistance at delivery: Doctor}
   \item{\code{MN19B}}{Assistance at delivery: Nurse / Midwife}
   \item{\code{MN19C}}{Assistance at delivery: Lady Health Visitor}
   \item{\code{MN19D}}{Assistance at delivery: Community Mid Wife}
   \item{\code{MN19F}}{Assistance at delivery: Traditional birth attendant}
   \item{\code{MN19H}}{Assistance at delivery: Relative / Friend}
   \item{\code{MN19X}}{Assistance at delivery: Other}
   \item{\code{MN19Y}}{Assistance at delivery: No one}
   \item{\code{MN19NR}}{Assistance at delivery: No response}
   \item{\code{MN20}}{Place of delivery}
   \item{\code{MN21}}{Delivery by caesarean section}
   \item{\code{MN22}}{Decision made to have caesarean section}
   \item{\code{MN23}}{After the birth, baby was put directly on the bare skin of mothers chest}
   \item{\code{MN24}}{Baby was wrapped up before being placed on mothers chest}
   \item{\code{MN25}}{Baby was dried or wiped soon after birth}
   \item{\code{MN26U}}{Time baby was bathed - unit}
   \item{\code{MN26N}}{Time baby was bathed - number}
   \item{\code{MN28}}{What was used to cut the cord}
   \item{\code{MN29}}{Instrument used to cut the cord boiled prior to use}
   \item{\code{MN30}}{Something applied to the cord after it was cut}
   \item{\code{MN31A}}{What was applied to the cord: Chlorhexidine}
   \item{\code{MN31B}}{What was applied to the cord: Other antiseptic}
   \item{\code{MN31C}}{What was applied to the cord: Mustard oil}
   \item{\code{MN31D}}{What was applied to the cord: Ash}
   \item{\code{MN31E}}{What was applied to the cord: Animal dung}
   \item{\code{MN31X}}{What was applied to the cord: Other}
   \item{\code{MN31Z}}{What was applied to the cord: DK / Don't remember}
   \item{\code{MN31NR}}{What was applied to the cord: No Response}
   \item{\code{MN32}}{Size of child at birth}
   \item{\code{MN33}}{Weighted at birth}
   \item{\code{MN34A}}{Weight from card or recall}
   \item{\code{MN34}}{Weight at birth (Kilograms)}
   \item{\code{MN35}}{Menstrual period returned since the birth of child}
   \item{\code{MN36}}{Ever breastfeed}
   \item{\code{MN37U}}{Time baby put to breast - unit}
   \item{\code{MN37N}}{Time baby put to breast - number}
   \item{\code{MN38}}{Within first 3 days after delivery, child given anything to drink other than breast milk}
   \item{\code{MN39A}}{Child given to drink: Milk (other than breast milk)}
   \item{\code{MN39B}}{Child given to drink: Plain water}
   \item{\code{MN39C}}{Child given to drink: Sugar or glucose water}
   \item{\code{MN39D}}{Child given to drink: Gripe water}
   \item{\code{MN39E}}{Child given to drink: Sugar - salt - water solution}
   \item{\code{MN39F}}{Child given to drink: Fruit juice}
   \item{\code{MN39G}}{Child given to drink: Infant formula}
   \item{\code{MN39H}}{Child given to drink: Tea / Infusions}
   \item{\code{MN39I}}{Child given to drink: Honey}
   \item{\code{MN39J}}{Child given to drink: Prescribed medicine}
   \item{\code{MN39K}}{Child given to drink: Rose Water}
   \item{\code{MN39X}}{Child given to drink: Other}
   \item{\code{MN39Y}}{Child given to drink: Not given anything to drink}
   \item{\code{MN39NR}}{Child given to drink: No response}
   \item{\code{MN40}}{Lady Health Worker}
   \item{\code{PN3U}}{Duration of staying in the health facility - unit}
   \item{\code{PN3N}}{Duration of staying in the health facility - number}
   \item{\code{PN4}}{Baby's health checked before leaving health facility}
   \item{\code{PN5}}{Mother's health checked before leaving health facility}
   \item{\code{PN5AA}}{Before leaving the facility, did anyone: Counsel for family planning}
   \item{\code{PN5AB}}{Before leaving the facility, did anyone: Tell about various family planning methods}
   \item{\code{PN5AC}}{Before leaving the facility, did anyone: Guide about from where you can get these methods}
   \item{\code{PN6}}{Baby's health checked after leaving health facility}
   \item{\code{PN8}}{Baby checked after the delivery was over}
   \item{\code{PN9}}{Mother checked after the delivery was over}
   \item{\code{PN9AA}}{Councelling for Family Planning}
   \item{\code{PN9AB}}{Family Planning Methods}
   \item{\code{PN9AC}}{From where you get these methods}
   \item{\code{PN10}}{Additional baby check after the delivery}
   \item{\code{PN11}}{Baby checked after the delivery}
   \item{\code{PN12}}{Number of times baby was checked}
   \item{\code{PN13U}}{How long after delivery did the first check of baby happen - unit}
   \item{\code{PN13N}}{How long after delivery did the first check of baby happen - number}
   \item{\code{PN14A}}{Person checking on baby's health: Doctor}
   \item{\code{PN14B}}{Person checking on baby's health: Nurse / Midwife}
   \item{\code{PN14C}}{Person checking on baby's health: Lady Health Visitor (LHV)}
   \item{\code{PN14D}}{Person checking on baby's health: Community Mid Wife}
   \item{\code{PN14F}}{Person checking on baby's health: Traditional birth attendant}
   \item{\code{PN14H}}{Person checking on baby's health: Relative / Friend}
   \item{\code{PN14X}}{Person checking on baby's health: Other}
   \item{\code{PN14NR}}{Person checking on baby's health: Missing}
   \item{\code{PN15}}{Place where baby's health was checked}
   \item{\code{PN17}}{Mother's health checked after leaving health facility}
   \item{\code{PN19}}{Mother's health checked after the delivery was over}
   \item{\code{PN20}}{Mother's health checked after the birth of the baby}
   \item{\code{PN21}}{Number of times mother's health was checked after the birth}
   \item{\code{PN22U}}{How long after delivery did the first check of mother happen - unit}
   \item{\code{PN22N}}{How long after delivery did the first check of mother happen - number}
   \item{\code{PN23A}}{Person checking on mother's health: Doctor}
   \item{\code{PN23B}}{Person checking on mother's health: Nurse / Midwife}
   \item{\code{PN23C}}{Person checking on mother's health: Lady Health Visitor (LHV)}
   \item{\code{PN23D}}{Person checking on mother's health: Community MID Wife}
   \item{\code{PN23F}}{Person checking on mother's health: Traditional birth attendant}
   \item{\code{PN23H}}{Person checking on mother's health: Relative / Friend}
   \item{\code{PN23X}}{Person checking on mother's health: Other}
   \item{\code{PN23NR}}{Person checking on mother's health: Missing}
   \item{\code{PN24}}{Place where mother's health was checked}
   \item{\code{PN25A}}{During two days after birth health care provider: Examine cord}
   \item{\code{PN25B}}{During two days after birth health care provider: Take temperature}
   \item{\code{PN25C}}{During two days after birth health care provider: Counsel on breastfeeding}
   \item{\code{PN27}}{During two days after birth health care provider observe child's breastfeeding}
   \item{\code{PN29}}{During two days after birth health care provider weighter child again}
   \item{\code{PN30}}{During two days after birth health care provider gave informtion on symtomps}
   \item{\code{CP1}}{Currently pregnant}
   \item{\code{CP2}}{Currently using a method to avoid pregnancy}
   \item{\code{CP3}}{Ever used a method to avoid pregnancy}
   \item{\code{CP4A}}{Current method: Female sterilization}
   \item{\code{CP4B}}{Current method: Male sterilization}
   \item{\code{CP4C}}{Current method: IUD}
   \item{\code{CP4D}}{Current method: Injectables / Injection}
   \item{\code{CP4E}}{Current method: Implants}
   \item{\code{CP4F}}{Current method: Pill}
   \item{\code{CP4G}}{Current method: Male condom}
   \item{\code{CP4H}}{Current method: Female condom}
   \item{\code{CP4I}}{Current method: Diaphragm}
   \item{\code{CP4K}}{Current method:Lactational amenorrhoea method (LAM)}
   \item{\code{CP4L}}{Current method: Periodic abstinence / Rhythm / Calendar}
   \item{\code{CP4M}}{Current method: Withdrawal}
   \item{\code{CP4X}}{Current method: Other}
   \item{\code{CP4NR}}{Current method: Missing}
   \item{\code{UN2}}{Wanted to get pregnant at the time}
   \item{\code{UN4}}{Wanted baby later or did not want more children}
   \item{\code{UN5}}{Would like to have another child (currently pregnant)}
   \item{\code{UN7}}{Would like to have another child (not currently pregnant)}
   \item{\code{UN8U}}{Desired waiting time (unit)}
   \item{\code{UN8N}}{Desired waiting time (number)}
   \item{\code{UN11}}{Able to get pregnant}
   \item{\code{UN12A}}{Reason: Infrequent sex / No sex}
   \item{\code{UN12B}}{Reason: Menopausal}
   \item{\code{UN12C}}{Reason: Never menstruated}
   \item{\code{UN12D}}{Reason: Hysterectomy}
   \item{\code{UN12E}}{Reason: Trying to get pregnant for 2 years without result}
   \item{\code{UN12F}}{Reason: Postpartum amenorrheic}
   \item{\code{UN12G}}{Reason: Breastfeeding}
   \item{\code{UN12H}}{Reason: Too old}
   \item{\code{UN12I}}{Reason: Fatalistic}
   \item{\code{UN12X}}{Reason: Other}
   \item{\code{UN12Z}}{Reason: Don't know}
   \item{\code{UN12NR}}{Reason: Missing}
   \item{\code{UN14U}}{Start of last menstrual period - unit}
   \item{\code{UN14N}}{Start of last menstrual period - number}
   \item{\code{UN16}}{Social activities, school or work days not attended due to menstruation}
   \item{\code{UN17}}{Availability of private place for washing during last menstrual period}
   \item{\code{UN18}}{Used any materials such as sanitary pads, tampons or cloth}
   \item{\code{UN19}}{Materials reusable}
   \item{\code{DV1A}}{If she goes out with out telling husband: wife beating justified}
   \item{\code{DV1B}}{If she neglects the children: wife beating justified}
   \item{\code{DV1C}}{If she argues with husband: wife beating justified}
   \item{\code{DV1D}}{If she refuses sex with husband: wife beating justified}
   \item{\code{DV1E}}{If she burns the food: wife beating justified}
   \item{\code{VT1}}{Something taken by using force or someone threatening to use force}
   \item{\code{VT2}}{Victimisation happened during the last 12 months}
   \item{\code{VT3}}{Number of times victimisation happened in the last year}
   \item{\code{VT5}}{The last time that this happened, was anything stolen}
   \item{\code{VT6}}{Person(s) had a weapon}
   \item{\code{VT7A}}{Offender(s) had a knife, a gun or something else used as a weapon: A knife}
   \item{\code{VT7B}}{Offender(s) had a knife, a gun or something else used as a weapon: A gun}
   \item{\code{VT7X}}{Offender(s) had a knife, a gun or something else used as a weapon: Other weapon}
   \item{\code{VT7NR}}{Offender(s) had a knife, a gun or something else used as a weapon: Missing}
   \item{\code{VT8}}{Incident reported to the police}
   \item{\code{VT9}}{Physically attacked}
   \item{\code{VT10}}{Attack happened during the last 12 months}
   \item{\code{VT11}}{Number of times attack happened in the last year}
   \item{\code{VT12}}{Place attack happened the last time}
   \item{\code{VT13}}{Number of people involved in committing the offence}
   \item{\code{VT14}}{At least one of the offender(s) recognized at the time of the incident}
   \item{\code{VT17}}{Person(s) had a weapon}
   \item{\code{VT18A}}{Offender(s) had a knife, a gun or something else used as a weapon: A knife}
   \item{\code{VT18B}}{Offender(s) had a knife, a gun or something else used as a weapon: A gun}
   \item{\code{VT18X}}{Offender(s) had a knife, a gun or something else used as a weapon: Other weapon}
   \item{\code{VT18NR}}{Offender(s) had a knife, a gun or something else used as a weapon: Missing}
   \item{\code{VT19}}{Incident reported to the police}
   \item{\code{VT20}}{Feeling safe walking alone in neighbourhood after dark}
   \item{\code{VT21}}{Feeling safe at home alone after dark}
   \item{\code{VT22A}}{In the past 12 months, felt discriminated: Ethnic or immigration origin}
   \item{\code{VT22B}}{In the past 12 months, felt discriminated: Gender}
   \item{\code{VT22C}}{In the past 12 months, felt discriminated: Sexual orientation}
   \item{\code{VT22D}}{In the past 12 months, felt discriminated: Age}
   \item{\code{VT22E}}{In the past 12 months, felt discriminated: Religion or belief}
   \item{\code{VT22F}}{In the past 12 months, felt discriminated: Disability}
   \item{\code{VT22X}}{In the past 12 months, felt discriminated: any other reason}
   \item{\code{HA1}}{Ever heard of HIV or AIDS}
   \item{\code{HA2}}{Can avoid HIV by having uninfected husband who has no other wife}
   \item{\code{HA3}}{Can get HIV from mosquito bites}
   \item{\code{HA4}}{Can avoid HIV by using a condom correctly every time}
   \item{\code{HA5}}{Can get HIV by sharing food with a person who has HIV}
   \item{\code{HA6}}{Can get HIV through supernatural means}
   \item{\code{HA7}}{Healthy-looking person may have HIV}
   \item{\code{HA8A}}{HIV from mother to child during pregnancy}
   \item{\code{HA8B}}{HIV from mother to child during delivery}
   \item{\code{HA8C}}{HIV from mother to child through breastfeeding}
   \item{\code{HA10}}{Aware of drugs used to reduce the risk of transmission to the baby}
   \item{\code{HA13A}}{Given information about HIV during antenatal care visit: HIV from mother}
   \item{\code{HA13B}}{Given information about HIV during antenatal care visit: Things to do}
   \item{\code{HA13C}}{Given information about HIV during antenatal care visit: Tested for HIV}
   \item{\code{HA13D}}{Given information about HIV during antenatal care visit: Offered a test}
   \item{\code{HA14}}{Tested for HIV as part of antenatal care}
   \item{\code{HA15}}{Received results from test during antenatal care}
   \item{\code{HA16}}{Received consultation after testing during antenatal care}
   \item{\code{HA18}}{Offered test for HIV before delivery}
   \item{\code{HA19}}{Tested for HIV during delivery}
   \item{\code{HA20}}{Received results from test during delivery}
   \item{\code{HA22}}{Tested for HIV since test during pregnancy}
   \item{\code{HA23}}{Most recent time of testing for HIV}
   \item{\code{HA24}}{Ever been tested for HIV}
   \item{\code{HA25}}{Most recent time of testing for HIV}
   \item{\code{HA26}}{Received results of HIV test}
   \item{\code{HA27}}{Know a place to get HIV test}
   \item{\code{HA28}}{Heard of test kit for HIV testing}
   \item{\code{HA29}}{Ever tested yourself using test kit}
   \item{\code{HA30}}{Would buy fresh vegetables from shopkeeper with AIDS virus}
   \item{\code{HA31}}{Children living with HIV should be allowed to attend school with other children}
   \item{\code{HA32}}{People hesitate to take an HIV test because they are afraid of how other people will react if the test result is positive}
   \item{\code{HA33}}{People talk badly about people living with HIV, or who are thought to be living with HIV}
   \item{\code{HA34}}{People living with HIV, or thought to be living with HIV, lose the respect of other people}
   \item{\code{HA35}}{Agreement with the following statement: I would be ashamed if someone in my family had HIV}
   \item{\code{HA36}}{Fear of getting HIV in contact with the saliva of a person living with HIV}
   \item{\code{MM1}}{Number of births to natural mother}
   \item{\code{MM3}}{Any brothers and sisters from the same mother that you have not mentioned}
   \item{\code{MM4}}{Any brothers or sisters who do not live with you that you have not mentioned}
   \item{\code{MM5}}{Any brothers or sisters who died that you have not mentioned}
   \item{\code{MM6}}{Any brothers or sisters born to your natural mother, but who have a different natural father}
   \item{\code{MM7}}{Total number of siblings listed in MM1}
   \item{\code{MM8}}{Total right}
   \item{\code{MM12}}{Number of preceding births}
   \item{\code{TA1}}{Ever tried cigarette smoking}
   \item{\code{TA2}}{Age when cigarette was smoked for the first time}
   \item{\code{TA3}}{Currently smoking cigarettes}
   \item{\code{TA4}}{Number of cigarettes smoked in the last 24 hours}
   \item{\code{TA5}}{Number of days when cigarettes were smoked in past month}
   \item{\code{TA6}}{Ever tried any smoked tobacco products other than cigarettes}
   \item{\code{TA7}}{Used any smoked tobacco products during the last month}
   \item{\code{TA8A}}{Type of smoked tobacco product: Cigars}
   \item{\code{TA8B}}{Type of smoked tobacco product: Water pipe}
   \item{\code{TA8C}}{Type of smoked tobacco product: Cigarillos}
   \item{\code{TA8D}}{Type of smoked tobacco product: Pipe / Shesha}
   \item{\code{TA8X}}{Type of smoked tobacco product: Other}
   \item{\code{TA8NR}}{Type of smoked tobacco product: No response}
   \item{\code{TA9}}{Number of days when tobacco products where smoked in past month}
   \item{\code{TA10}}{Ever tried any form of smokeless tobacco products}
   \item{\code{TA11}}{Used any smokeless tobacco products during the last month}
   \item{\code{TA12A}}{Type of smokeless tobacco product used: Chewing tobacco}
   \item{\code{TA12B}}{Type of smokeless tobacco product used: Snuff}
   \item{\code{TA12D}}{Type of smokeless tobacco product used: Pan with Tobacco}
   \item{\code{TA12E}}{Type of smokeless tobacco product used: GUTKA}
   \item{\code{TA12F}}{Type of smokeless tobacco product used: NASWAR}
   \item{\code{TA12G}}{Type of smokeless tobacco product used: MAWA TOBACCO}
   \item{\code{TA12H}}{Type of smokeless tobacco product used: NAAS AND MAN PORI}
   \item{\code{TA12X}}{Type of smokeless tobacco product used: Other}
   \item{\code{TA12NR}}{Type of smokeless tobacco product used: No response}
   \item{\code{TA13}}{Number of days when smokeless tobacco products where used in past month}
   \item{\code{LS1}}{Estimation of overall happiness}
   \item{\code{LS2}}{Satisfaction with ladder step}
   \item{\code{LS3}}{Life satisfaction in comparison with last year}
   \item{\code{LS4}}{Life satisfaction expectation one year from now}
   \item{\code{HH4}}{Supervisor number}
   \item{\code{HH6}}{Area}
   \item{\code{WDOB}}{Date of birth of woman (CMC)}
   \item{\code{WDOM}}{Date of marriage (CMC)}
   \item{\code{WAGEM}}{Age at first marriage}
   \item{\code{WDOBFC}}{Date of birth of first child (CMC)}
   \item{\code{WDOBLC}}{Date of birth of last child (CMC)}
   \item{\code{MSTATUS}}{Marital status}
   \item{\code{CEB}}{Children ever born}
   \item{\code{CSURV}}{Children surviving}
   \item{\code{CDEAD}}{Children dead}
   \item{\code{wscore}}{Combined wealth score}
   \item{\code{windex5}}{Wealth index quintile}
   \item{\code{windex10}}{Percentile Group of com1}
   \item{\code{wscoreu}}{Urban wealth score}
   \item{\code{windex5u}}{Urban wealth index quintile}
   \item{\code{windex10u}}{Percentile Group of urb1}
   \item{\code{wscorer}}{Rural wealth score}
   \item{\code{windex5r}}{Rural wealth index quintile}
   \item{\code{windex10r}}{Percentile Group of rur1}
   \item{\code{division}}{Division}
   \item{\code{majorcity}}{majorcity}
   \item{\code{hh6r}}{Area}
   \item{\code{suburban}}{ }
   \item{\code{hh7}}{Districts}
   \item{\code{disability}}{Functional difficulties (age 18-49 years)}
   \item{\code{welevel}}{Education}
   \item{\code{insurance}}{Health insurance}
   \item{\code{BH4D_first}}{Day of birth}
   \item{\code{BH4M_first}}{Month of birth}
   \item{\code{BH4Y_first}}{Year of birth}
   \item{\code{BH6_first}}{Age of child}
   \item{\code{BH3_last}}{Sex of newborns}
   \item{\code{BH4D_last}}{Day of birth}
   \item{\code{BH4M_last}}{Month of birth}
   \item{\code{BH4Y_last}}{Year of birth}
   \item{\code{BH6_last}}{Age of child}
   \item{\code{wmweight}}{Woman's sample weight}
   \item{\code{stratum}}{stratum}
   \item{\code{psu}}{psu}
  }}
\usage{
data(PakPMICS2018wm)
}
\description{
\code{PakPMICS2018wm} data of Child questionnaire, MICS Punjab, 2014.
}
\examples{
data(PakPMICS2018wm)
}
\references{
\enumerate{
  \item UNICEF MICS (\url{http://www.mics.unicef.org/surveys}).
 }
}
\author{
\enumerate{
  \item{Muhammad Yaseen } {(\email{myaseen208@gmail.com})}
  \item{Muhammad Usman } {(\email{usmann75@hotmail.com})}
  }
}
\keyword{datasets}
